feat: Update code

+++ contracts/ArrayShift.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.14;
 
 contract ArrayShift {
-    uint[] public arr;
+    uint256[] public arr;
 
     function example() public {
         arr = [1, 2, 3];
@@ -14,14 +14,13 @@ contract ArrayShift {
     // [1, 2, 3] -> remove(1) -> [1, 3, 3] -> [1, 3]
     // [1, 2, 3, 4, 5, 6] -> remove(2) -> [1, 2, 4, 5, 6, 6] -> [1, 2, 4, 5, 6]
     // [1] -> remove(0) -> [1] -> []
-    function remove(uint _index) public {
+    function remove(uint256 _index) public {
         require(_index < arr.length, "index out of bound");
-        for (uint i = _index; i < arr.length - 1; i++) {
+        for (uint256 i = _index; i < arr.length - 1; i++) {
             arr[i] = arr[i + 1];
         }
         arr.pop();
     }
-
 
     function test() external {
         arr = [1, 2, 3, 4, 5];
@@ -39,4 +38,3 @@ contract ArrayShift {
         assert(arr.length == 0);
     }
 }
-
